refactor(grid): add tree plugin
