+++ packages/vue/src/grid/src/checkbox/src/methods.ts
@@ -2,6 +2,7 @@ import { hasCheckField, hasNoCheckField } from './handleSelectRow'
 import { hasCheckFieldNoStrictly, hasNoCheckFieldNoStrictly, setSelectionNoStrictly } from './setAllSelection'
 import { getTableRowKey } from '../../table/src/strategy'
 import { emitEvent } from '@opentiny/vue-renderless/grid/utils'
+import { isArray, set, get, eachTree, find, toStringJSON, toArray } from '@opentiny/vue-renderless/grid/static/'
 
 export default {
   // 处理默认勾选
@@ -162,5 +163,105 @@ export default {
     })
 
     return this.$nextTick()
+  },
+  initMultipleHistory() {
+    const { isMultipleHistory, toolBarVm } = this.$grid
+    const {
+      settingOpts: { storageKey },
+      id: toolbarId
+    } = toolBarVm
+    let remoteSelectedMethod = toolBarVm.setting.multipleHistory.remoteSelectedMethod
+    let remoteSelectedPromise
+
+    if (
+      isMultipleHistory &&
+      toolBarVm &&
+      toolBarVm.setting &&
+      toolBarVm.setting.multipleHistory &&
+      remoteSelectedMethod
+    ) {
+      if (typeof remoteSelectedMethod === 'function') {
+        remoteSelectedPromise = remoteSelectedMethod()
+
+        if (typeof remoteSelectedPromise.then === 'function') {
+          remoteSelectedPromise.then((storeStr) => {
+            let storeObj = toStringJSON(storeStr)
+            storeObj = (storeObj && storeObj[storageKey]) || null
+            storeObj = (storeObj || {})[toolbarId] || {}
+            const { columns, pageSize } = storeObj
+            toolBarVm.applySettings({ columns, pageSize })
+          })
+        }
+      }
+    }
+  },
+  // 显示多选工具栏
+  showSelectToolbar() {
+    let {
+      $grid: { selectToolbar, showHeader },
+      selectToolbarStore
+    } = this
+    if (selectToolbar && showHeader) {
+      selectToolbarStore.visible = false
+      let selectColumn = find(this.visibleColumn, (item) => item.type === 'selection')
+      let selected = this.getSelectRecords()
+      let position = typeof selectToolbar === 'object' ? selectToolbar.position : ''
+      if (selectColumn && selected && selected.length) {
+        let selectTh = this.$el.querySelector('th.tiny-grid-header__column.col__selection')
+        let headerWrapper = this.$el.querySelector('.tiny-grid>.tiny-grid__header-wrapper')
+        let tr = selectTh.parentNode
+        let thArr = toArray(tr.childNodes)
+        let range = document.createRange()
+        let rangeBoundingRect
+        let headerBoundingRect = headerWrapper.getBoundingClientRect()
+        let layout = { width: 0, height: 0, left: 0, top: 0, zIndex: 1 }
+        let adjust = 1
+        if (selectColumn.fixed === 'right') {
+          range.setStart(tr, thArr.indexOf(selectTh))
+          range.setEnd(tr, thArr.length)
+          rangeBoundingRect = range.getBoundingClientRect()
+          layout.left = `${adjust}px`
+        } else {
+          range.setStart(tr, 0)
+          range.setEnd(tr, thArr.indexOf(selectTh) + 1)
+          rangeBoundingRect = range.getBoundingClientRect()
+          layout.left = `${rangeBoundingRect.width + adjust}px`
+        }
+        layout.width = `${headerBoundingRect.width - rangeBoundingRect.width - 2 * adjust}px`
+        if (!selectColumn.fixed && position === 'left') {
+          range = document.createRange()
+          range.setStart(tr, 0)
+          range.setEnd(tr, thArr.indexOf(selectTh))
+          rangeBoundingRect = range.getBoundingClientRect()
+          layout.left = `${adjust}px`
+          layout.width = `${rangeBoundingRect.width - 2 * adjust}px`
+        }
+        layout.top = `${headerBoundingRect.height - rangeBoundingRect.height + adjust}px`
+        layout.height = `${rangeBoundingRect.height - 2 * adjust}px`
+        return this.$nextTick().then(() => {
+          selectToolbarStore.layout = layout
+          selectToolbarStore.visible = true
+        })
+      }
+    }
+    return this.$nextTick()
+  },
+  // 切换多选工具栏的显示
+  toggleSelectToolbarVisible() {
+    this.selectToolbarStore.visible = !this.selectToolbarStore.visible
+    return this.$nextTick()
+  },
+  // 在空数据时Selection列表头复选框禁用，headerAutoDisabled设置为false就会和旧版本兼容
+  handleSelectionHeader() {
+    const { tableFullData, visibleColumn, selectConfig = {} } = this
+    const { headerAutoDisabled } = selectConfig
+    const selectionColumn = visibleColumn.find((column) => column.type === 'selection')
+    if (
+      (typeof headerAutoDisabled === 'undefined' || (typeof headerAutoDisabled === 'boolean' && headerAutoDisabled)) &&
+      !tableFullData.length &&
+      selectionColumn
+    ) {
+      this.headerCheckDisabled = true
+    }
   }
 }
